requirements_comparison.txt
# Additional requirements for model comparison script
# Install with: pip install -r requirements_comparison.txt

# Data analysis and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Image processing
scikit-image>=0.18.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Deep learning (if not already installed)
torch>=1.9.0
torchvision>=0.10.0

# Utilities
tqdm>=4.62.0